// src/app/models/orderdocument.model.ts
export interface Orderdocument {
    orderDocumentId: string;
    documentType: string;
    orderQtyId: number;
	userId: number;
	userName: string;
	role: string;
    manufacturerId: number;
    mfgName: string;
    productId: number;
    productName: string;
    productDescription: string;
    casNumber: string;
    quantity: number;
    createdAt: string;
}  
	